# boost performance of normalization