as benchmark data shown below, we can get **43%** performance improvement **at most**.
all unit tests are passed successfully.
BTW: we can also consolidate these commonly used preprocessing methods into a standalone 
file for maximizing code reusability.

```

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19044.3693/21H2/November2021Update)
.NET SDK 8.0.100
  [Host]     : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-YOWNUE : .NET 7.0.14 (7.0.1423.51910), X64 RyuJIT AVX2
  Job-GQLGBL : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

IterationCount=10  LaunchCount=1  WarmupCount=1  

```
| Method          | Runtime  | FileName   | Mean      | Error     | StdDev   | Ratio | RatioSD | Gen0     | Gen1     | Gen2     | Allocated | Alloc Ratio |
|---------------- |--------- |----------- |----------:|----------:|---------:|------:|--------:|---------:|---------:|---------:|----------:|------------:|
| **NormalizeBySdcb** | **.NET 7.0** | **test_1.jpg** | **128.01 μs** |  **7.205 μs** | **4.766 μs** |  **1.00** |    **0.00** |   **0.1221** |        **-** |        **-** |     **816 B** |        **1.00** |
| UnsafeWayByAven | .NET 7.0 | test_1.jpg |  82.20 μs |  1.585 μs | 1.048 μs |  0.64 |    0.02 | 100.7080 | 100.2197 | 100.2197 |  389504 B |      477.33 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| NormalizeBySdcb | .NET 8.0 | test_1.jpg | 125.09 μs |  9.353 μs | 6.186 μs |  1.00 |    0.00 |   0.1221 |        - |        - |     816 B |        1.00 |
| UnsafeWayByAven | .NET 8.0 | test_1.jpg |  71.63 μs |  0.932 μs | 0.617 μs |  **0.57** |    0.03 |  98.3887 |  98.0225 |  98.0225 |  389554 B |      477.39 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| **NormalizeBySdcb** | **.NET 7.0** | **test_2.jpg** | **193.06 μs** |  **4.904 μs** | **3.244 μs** |  **1.00** |    **0.00** |        **-** |        **-** |        **-** |     **816 B** |        **1.00** |
| UnsafeWayByAven | .NET 7.0 | test_2.jpg | 148.16 μs |  2.267 μs | 1.349 μs |  0.77 |    0.01 | 167.4805 | 166.5039 | 166.5039 |  685119 B |      839.61 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| NormalizeBySdcb | .NET 8.0 | test_2.jpg | 192.66 μs |  4.812 μs | 2.863 μs |  1.00 |    0.00 |        - |        - |        - |     816 B |        1.00 |
| UnsafeWayByAven | .NET 8.0 | test_2.jpg | 133.71 μs |  2.004 μs | 1.326 μs |  0.69 |    0.01 | 167.4805 | 166.5039 | 166.5039 |  685119 B |      839.61 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| **NormalizeBySdcb** | **.NET 7.0** | **test_3.jpg** | **222.86 μs** |  **2.203 μs** | **1.311 μs** |  **1.00** |    **0.00** |        **-** |        **-** |        **-** |     **816 B** |        **1.00** |
| UnsafeWayByAven | .NET 7.0 | test_3.jpg | 172.17 μs |  2.277 μs | 1.506 μs |  0.77 |    0.01 | 197.7539 | 196.7773 | 196.7773 |  792868 B |      971.65 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| NormalizeBySdcb | .NET 8.0 | test_3.jpg | 223.14 μs |  4.353 μs | 2.879 μs |  1.00 |    0.00 |        - |        - |        - |     816 B |        1.00 |
| UnsafeWayByAven | .NET 8.0 | test_3.jpg | 156.06 μs |  4.578 μs | 3.028 μs |  0.70 |    0.02 | 195.3125 | 194.3359 | 194.3359 |  792833 B |      971.61 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| **NormalizeBySdcb** | **.NET 7.0** | **test_4.jpg** | **397.70 μs** |  **9.467 μs** | **6.262 μs** |  **1.00** |    **0.00** |        **-** |        **-** |        **-** |     **816 B** |        **1.00** |
| UnsafeWayByAven | .NET 7.0 | test_4.jpg | 324.22 μs |  8.080 μs | 4.808 μs |  0.82 |    0.02 | 280.2734 | 279.2969 | 279.2969 | 1429175 B |    1,751.44 |
|                 |          |            |           |           |          |       |         |          |          |          |           |             |
| NormalizeBySdcb | .NET 8.0 | test_4.jpg | 399.24 μs |  11.727 μs | 7.757 μs |  1.00 |    0.00 |        - |        - |        - |     816 B |        1.00 |
| UnsafeWayByAven | .NET 8.0 | test_4.jpg | 299.93 μs | 8.187 μs | 5.415 μs|  0.75 |    0.01 | 243.1641 | 242.1875 | 242.1875 | 1428748 B |    1,750.92 |

benchmark code

```
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Jobs;
using OpenCvSharp;
using System.Runtime.CompilerServices;

namespace Sdcb.OpenVINO.Benchmarks
{
    [SimpleJob(runtimeMoniker: RuntimeMoniker.Net70, launchCount: 1, warmupCount: 1, iterationCount: 10)]
    [SimpleJob(runtimeMoniker: RuntimeMoniker.Net80, launchCount: 1, warmupCount: 1, iterationCount: 10)]
    [MemoryDiagnoser]
    [HtmlExporter]
    [MarkdownExporterAttribute.GitHub]
    public class MatMeanStd
    {
        private static readonly Vec3f meanValues = new(
            255.0f * 0.485f,
            255.0f * 0.456f,
            255.0f * 0.406f);

        private static readonly Vec3f stdValues = new(
            1 / (255.0f * 0.229f),
            1 / (255.0f * 0.224f),
            1 / (255.0f * 0.225f));

        static readonly string baseDir = AppContext.BaseDirectory;
        private Mat mat;

        [Params("test_1.jpg", "test_2.jpg", "test_3.jpg", "test_4.jpg")]
        public string FileName { get; set; }

        [GlobalSetup]
        public void GlobalSetup()
        {
            this.mat = new Mat(Path.Combine(baseDir, "images", this.FileName));

            Cv2.CvtColor(this.mat, this.mat, ColorConversionCodes.BGR2RGB);
        }

        [Benchmark(Description = "NormalizeBySdcb", Baseline = true)]
        public void NormalizeBySdcb()
        {
            NormalizeSdcb(this.mat);
        }  

        [Benchmark(Description = "UnsafeWayByAven")]
        public unsafe void UnsafeWayByAven()
        {
            UnsafeWayAven(this.mat);
        }

        private static unsafe Mat UnsafeWayAven(Mat src)
        {
            if (src.Type() != MatType.CV_32SC3)
            {
                src.ConvertTo(src, MatType.CV_32FC3);
            }

            int height = src.Height;
            int width = src.Width;
            int channel = src.Channels();
            float[] dstFloat = new float[width * height * channel];

            src.GetArray(out Vec3f[]? pixels);
            ref float srcFloat = ref Unsafe.As<Vec3f, float>(ref pixels[0]);

            fixed (float* pSrc = &srcFloat)
            fixed (float* pDst = &dstFloat[0])
            {
                Unsafe.CopyBlockUnaligned(pDst, pSrc, (uint)dstFloat.Length * sizeof(float));

                float* pointer = pDst;

                for (int i = 0; i < height; i++)
                {
                    for (int j = 0; j < width; j++)
                    {
                        *pointer = (*pointer - meanValues.Item0) * stdValues.Item0;
                        pointer += 1;
                        *pointer = (*pointer - meanValues.Item1) * stdValues.Item1;
                        pointer += 1;
                        *pointer = (*pointer - meanValues.Item2) * stdValues.Item2;
                        pointer += 1;
                    }
                }
            }

            return new Mat(height, width, MatType.CV_32FC(channel), dstFloat);
        }

        private static Mat NormalizeSdcb(Mat src)
        {
            using Mat normalized = new();
            src.ConvertTo(normalized, MatType.CV_32FC3, 1.0 / 255);
            Mat[] bgr = normalized.Split();
            float[] scales = new[] { 1 / 0.229f, 1 / 0.224f, 1 / 0.225f };
            float[] means = new[] { 0.485f, 0.456f, 0.406f };
            for (int i = 0; i < bgr.Length; ++i)
            {
                bgr[i].ConvertTo(bgr[i], MatType.CV_32FC1, 1.0 * scales[i], (0.0 - means[i]) * scales[i]);
            }

            Mat dest = new();
            Cv2.Merge(bgr, dest);

            foreach (Mat channel in bgr)
            {
                channel.Dispose();
            }

            return dest;
        }

        ~MatMeanStd()
        {
            mat?.Dispose();
        }
    }
}

```