@sdcb 2023-12-17
Wow, that's pretty amazing perf boost, can I tell the reason is: it will automatically using SIMD?