@AvenSun 2023-12-17
SIMD operations can be parallelized at the hardware level. 
I'm afraid this implantation has nothing to do with SIMD.
the improvement should be due to using **pointer** and **Unsafe.As** and **Unsafe.CopyBlockUnaligned**.