@sdcb 2023-12-19
@AvenSun How can you tell a piece of code can get better performance? Some code can benifits from C++, but some other code can benifits by raw C# pointer? just curious