@AvenSun 2023-12-19
hi @sdcb 
that's a good question.
for evaluating the performance of code snippets, benchmarks are a well-known and effective method.
If we can ensure that the methods involved maintain high-performance implementations, 
the overall performance of an application should be higher. 
I think it is unnecessary for us to ascertain whether this is attributed to native code or managed code.
from a developer's perspective, we can employ specialized tools to thoroughly examine performance information.
for instance, the .NET framework offers tools such as dotnet-monitor, dotnet-trace, dotnet-counters, and others for in-depth performance analysis.